+++ src/propagators/j2.jl
@@ -285,7 +285,7 @@ end
     ) where {
         Tjd <: Number,
         Tv <: AbstractVector
-    } -> KeplerianElements{MeanAnomaly, Float64, T}, SMatrix{6, 6, T}
+    } -> KeplerianElements{MeanAnomaly, Float64, T}, SMatrix{6, 6, T}, NamedTuple
 
 Fit a set of mean Keplerian elements for the J2 orbit propagator using the osculating
 elements represented by a set of position vectors `vr_i` [m] and a set of velocity vectors
@@ -342,6 +342,18 @@ elements represented by a set of position vectors `vr_i` [m] and a set of veloci
 
 - `KeplerianElements{MeanAnomaly, Float64, T}`: Fitted Keplerian elements.
 - `SMatrix{6, 6, T}`: Final covariance matrix of the least-square algorithm.
+- `NamedTuple`: Statistics of the least-square algorithm with the following fields:
+    - `converged::Bool`: `true` if the iterations stopped because the residue was lower
+        than `atol` or its relative variation was lower than `rtol`, or `false` if they
+        stopped by reaching `max_iterations`.
+    - `iterations::Int`: Number of iterations performed.
+    - `position_rmse::T`: RMSE of the position residue in the last iteration [m].
+    - `velocity_rmse::T`: RMSE of the velocity residue in the last iteration [m / s].
+    - `total_rmse::T`: Weighted RMSE of the residue in the last iteration.
+
+    The statistics refer to the fitting of the mean elements. If their epoch is updated
+    afterward to match `mean_elements_epoch`, the statistics of that update are not
+    returned.
 
 # Examples
 
@@ -373,13 +385,11 @@ julia> vjd = [
            2.4600282560230047e6
        ];
 
-julia> orb, P = fit_j2_mean_elements(vjd, vr_i, vv_i)
+julia> orb, P, stats = fit_j2_mean_elements(vjd, vr_i, vv_i);
 ACTION:   Fitting the mean elements for the J2 propagator.
            Iteration        Position RMSE        Velocity RMSE           Total RMSE       RMSE Variation
                                      [km]             [km / s]                  [ ]
 PROGRESS:          4               4.3413           0.00540076              4341.31          0.000476014 %
-
-(KeplerianElements{MeanAnomaly, Float64, Float64}: Epoch = 2.46003e6 (2023-03-24T18:08:40.388), [0.16604846210532387 0.06643574115992787 … -3.855413667214467e-5 0.0001240322538280385; 0.0664357411575934 0.2663344822298673 … -1.7943611754171365e-5 -1.9567957022170893e-5; … ; -3.8554136673065196e-5 -1.7943611755807416e-5 … 4.397198441146852e-7 -8.092704668569453e-8; 0.0001240322538286946 -1.9567957020172356e-5 … -8.092704668536403e-8 1.2451922426408166e-7])
 
 julia> orb
 KeplerianElements{MeanAnomaly, Float64, Float64}:
@@ -390,6 +400,9 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
   RA of Asc. Node   : 162.1773906°
   Arg. of Periapsis : 101.2857729°
   Mean Anomaly      : 258.8172342°
+
+julia> stats
+(converged = true, iterations = 4, position_rmse = 4341.303234165598, velocity_rmse = 5.400761345997574, total_rmse = 4341.306593549895)
 ```
 """
 function fit_j2_mean_elements(
@@ -419,7 +432,7 @@ end
         Tepoch <: Number,
         Tjd <: Number,
         Tv <: AbstractVector
-    } -> KeplerianElements{MeanAnomaly, Tepoch, T}, SMatrix{6, 6, T}
+    } -> KeplerianElements{MeanAnomaly, Tepoch, T}, SMatrix{6, 6, T}, NamedTuple
 
 Fit a set of mean Keplerian elements for the J2 orbit propagator `j2d` using the osculating
 elements represented by a set of position vectors `vr_i` [m] and a set of velocity vectors
@@ -472,6 +485,18 @@ elements represented by a set of position vectors `vr_i` [m] and a set of veloci
 
 - `KeplerianElements{MeanAnomaly, Tepoch, T}`: Fitted Keplerian elements.
 - `SMatrix{6, 6, T}`: Final covariance matrix of the least-square algorithm.
+- `NamedTuple`: Statistics of the least-square algorithm with the following fields:
+    - `converged::Bool`: `true` if the iterations stopped because the residue was lower
+        than `atol` or its relative variation was lower than `rtol`, or `false` if they
+        stopped by reaching `max_iterations`.
+    - `iterations::Int`: Number of iterations performed.
+    - `position_rmse::T`: RMSE of the position residue in the last iteration [m].
+    - `velocity_rmse::T`: RMSE of the velocity residue in the last iteration [m / s].
+    - `total_rmse::T`: Weighted RMSE of the residue in the last iteration.
+
+    The statistics refer to the fitting of the mean elements. If their epoch is updated
+    afterward to match `mean_elements_epoch`, the statistics of that update are not
+    returned.
 
 # Examples
 
@@ -506,13 +531,11 @@ julia> vjd = [
            2.4600282560230047e6
        ];
 
-julia> orb, P = fit_j2_mean_elements!(j2d, vjd, vr_i, vv_i)
+julia> orb, P, stats = fit_j2_mean_elements!(j2d, vjd, vr_i, vv_i);
 ACTION:   Fitting the mean elements for the J2 propagator.
            Iteration        Position RMSE        Velocity RMSE           Total RMSE       RMSE Variation
                                      [km]             [km / s]                  [ ]
 PROGRESS:          4               4.3413           0.00540076              4341.31          0.000476014 %
-
-(KeplerianElements{MeanAnomaly, Float64, Float64}: Epoch = 2.46003e6 (2023-03-24T18:08:40.388), [0.16604846210532387 0.06643574115992787 … -3.855413667214467e-5 0.0001240322538280385; 0.0664357411575934 0.2663344822298673 … -1.7943611754171365e-5 -1.9567957022170893e-5; … ; -3.8554136673065196e-5 -1.7943611755807416e-5 … 4.397198441146852e-7 -8.092704668569453e-8; 0.0001240322538286946 -1.9567957020172356e-5 … -8.092704668536403e-8 1.2451922426408166e-7])
 
 julia> orb
 KeplerianElements{MeanAnomaly, Float64, Float64}:
@@ -523,6 +546,9 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
   RA of Asc. Node   : 162.1773906°
   Arg. of Periapsis : 101.2857729°
   Mean Anomaly      : 258.8172342°
+
+julia> stats
+(converged = true, iterations = 4, position_rmse = 4341.303234165598, velocity_rmse = 5.400761345997574, total_rmse = 4341.306593549895)
 ```
 """
 function fit_j2_mean_elements!(